feat(수익률을 계산합니다.)

+++ src/test/java/lotto/domain/MatchCriteriaTest.java
@@ -25,9 +25,9 @@ class MatchCriteriaTest {
         MatchCriteria three = MatchCriteria.THREE;
         three.match();
         //when
-        Long result = three.calc();
+        double result = three.calc();
         //then
-        assertThat(result).isEqualTo(5000L);
+        assertThat(result).isEqualTo(5000);
     }
 
 }
